requirements.txt
Flask
flask-cors
numpy
opencv-python
Pillow
tensorflow
gunicorn
h5py
scipy